Update caching in model config and predict function refactor
In this commit, the model configuration is updated for adding a caching directory for AutoConfig when it is initialized. Also, various predict_iteration functions have been refactored to inference_iteration to better denote their focus on performing inference on a data sample. Additionally, parameters, imports and defaults were updated in several modules to accommodate fine tuning. And several other minor modifications were made throughout the codebase to improve readability and functionality.