Add gRPC for serving alongside with http
Added endpoints and logic in grpc_server.py to handle incoming gRPC requests. Refactored beam_server.py to create a parent class. The HTTP-specific logic was moved into http_server.py and the common logic was left in the parent class. Added grpcio-tools to the requirements file to facilitate gRPC usage. The gRPC client code was added in the GRPCClient class in grpc_client.py. Updated __init__.py to export GRPCClient and GRPCServer classes. This change was necessary to provide more flexibility to the users when serving models, as some might prefer gRPC over HTTP due to its advantages such as efficiency and better support for streaming requests.